rc: move files around to enforce execution order
Some files where ordered "by luck". Ensure they are executed when we
expect them:

 - `helpers` is needed early (not really, but the intent is here)
 - `hostname` is needed very early as correct hostname is needed early
 - `path` is theoretically needed early

Then, we need to setup `run` early (but we need `hostname` to be set
correctly).

`terminfo` is important but mainly for interactive stuff. So, don't set
it too early.

`title` is like helpers but being more specific, we can set it later.

`locale` is important but not before the others.