workon: don't use a coproc for Docker containers
It was pretty cool, but not needed. We can enter the other namespaces
as well, unshare the mount namespace and do stuff. At the end, the
security is the same. We have the same privileges than the processes
in the container: they could nsenter our namespace. From there, they
have full access of home and they could also retrieve the mount point
of the root filesystem and mount it again.

Note that Docker prevents most of it by default (unless this is a
privileged container).