prompt: detect "nix shell" environment
No environment variable unfortunately, so if we are no "IN_NIX_SHELL"
but we have /nix/store stuff in PATH, let's just consider we are in a
"nix shell".

However, we are also unable to detect pure/impure this way, so the
environment will always be impure since we will pull all paths in
zshenv.

See https://github.com/NixOS/nix/issues/3862 for more context.